Update index.md
The Article got Partial pasted into a discussion in https://github.com/microsoft/TypeScript/issues/27957#issuecomment-893171119 and i found out that it makes false claims that confuses new users. So i corrected the relevant parts as the Users do not understand that we migrate forward to ESM and not ship old none needed CJS if it is not needed.